refactor(test): extract test_utils into module directory with shared graph fixtures
- Split test_utils.rs into test_utils/{graph_fixtures,graph_lookup,marginal}.rs for navigability
- Move TestNode/TestEdge from graph/__tests__/graph.rs into shared graph_fixtures, eliminating cross-module test type access via crate::graph::__tests__::graph::tests
- Make graph module #[cfg(test)] (contains no production code), prefix test functions with test_graph_/test_edge_
- Update test inventory paths and fixture/helper references
